security/requirements.txt
fastapi==0.115.12  # Updated to latest version
uvicorn[standard]==0.34.0
pydantic==2.11.4  # Updated to latest version
sqlalchemy==2.0.38
# Removing python-jose due to vulnerabilities CVE-2024-33664 and CVE-2024-33663
# # python-jose[cryptography]==3.3.0  # Removed due to vulnerabilities CVE-2024-33664 and CVE-2024-33663
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.20  # Updated to latest version to fix CVE-2024-539810  # Updated to latest version to fix CVE-2024-53981
cryptography==46.0.0  # Updated to latest version to fix CVE-2024-26130, CVE-2024-12797, CVE-2024-6119
pyopenssl==24.0.0
certifi==2024.10.5  # Updated to latest version
oauthlib==3.2.2
pyotp==2.9.0
qrcode==7.4.2
slowapi==0.1.9
pyjwt==2.8.0  # Using a secure version that doesn't have CVE-2024-53861  # Downgraded to fix CVE-2024-53861
bandit==1.7.8
safety==3.0.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
redis==5.0.4  # Updated to latest version
httpx==0.27.0
itsdangerous==2.2.0
secure==0.3.0
email-validator==2.1.1
aiofiles==23.2.1  # Updated to latest version
loguru==0.7.2
prometheus-client==0.20.0
structlog==24.1.0
